[ci-app] move test to its own file
